generated_tools/2026-03-02/claude_doc_autoassembler/requirements.txt
jinja2
python-docx
fpdf
requests
